m2m/pyproject.toml
[tool.poetry]
name = "cli"
version = "0.1.0"
description = ""
authors = ["Todd Greenwood <tgreenwood@redzone.co>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
typer = "^0.12.3"
auth0-python = "^4.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
cli = "cli.console:app"
